Restructure to put most of header functionality in .cc and ensure Quadrotor cannot transmorgrify for symbolic expression
