+++ examples/quadrotor/test/quadrotor_dynamics_test.cc
@@ -232,7 +232,7 @@ TEST_F(QuadrotorTest, ToAutoDiff) {
 
 TEST_F(QuadrotorTest, ToSymbolic) {
   const QuadrotorPlant<double> plant;
-  EXPECT_TRUE(is_symbolic_convertible(plant));
+  EXPECT_FALSE(is_symbolic_convertible(plant));
 }
 
 }  // namespace
